requirements.txt
ddt<=1.7.2
django<=5.1
djangorestframework<=3.15.2
django-hashids<=0.7.0
pytest<=8.3.3
